Abstract in page submenu into separate component
